pyproject.toml
[project]
name = "mcp-remote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "mcp[cli]>=1.9.3",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.7.7",
    "httpx>=0.25.0",
    "PyJWT>=2.8.0",
    "scalekit-sdk-python>=2.2.0",
]

[project.scripts]
mcp-server = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
